+++ Source/Infra/EKContentView.swift
@@ -113,7 +113,7 @@ class EKContentView: UIView {
         // Define a spacer to catch top / bottom offsets
         var spacerView: UIView!
         let safeAreaInsets = EKWindowProvider.safeAreaInsets
-        let overrideSafeArea = attributes.positionConstraints.safeArea.isOverriden
+        let overrideSafeArea = attributes.positionConstraints.safeArea.isOverridden
         
         if !overrideSafeArea && safeAreaInsets.hasVerticalInsets && !attributes.position.isCenter {
             spacerView = UIView()