+++ CHANGELOG.md
@@ -1,6 +1,58 @@
 # Change Log
 Any notable changes to this project will be documented in this file.
 
+## 0.3.0
+
+### Bug Fixes
+
+#### Typos
+`EKAttributes.PositionConstraints.SafeArea.isOverriden` to `EKAttributes.PositionConstraints.SafeArea.isOverridden`
+`EKAttributes.PositionConstraints.SafeArea.overriden` to `EKAttributes.PositionConstraints.SafeArea.overridden`
+
+### Features
+
+#### Added an entry transform feature (ALPHA FEATURE)
+
+Developers are able to transform an entry to another entry using the same attributes.
+
+```Swift
+let view = UIView()
+// Customize
+SwiftEntryKit.transform(to: view)
+```
+
+#### Rating Popup 
+
+Added a rating popup (See custom presets).
+See also: `EKRatingMessage` and `EKRatingMessageView`
+
+
+### Other Changes:
+
+#### Image-less popups
+
+`EKPopUpMessage` can be image-less by simply setting `themeImage` to `nil` (or leaving its default value as is).
+
+```Swift
+/** Popup theme image */
+public struct ThemeImage {
+
+    /** Position of the theme image */
+    public enum Position {
+        case topToTop(offset: CGFloat)
+        case centerToTop(offset: CGFloat)
+    }
+
+    /** The content of the image */
+    public var image: EKProperty.ImageContent
+
+    /** The psotion of the image */
+    public var position: Position
+}
+
+public init(themeImage: ThemeImage? = default, title: EKProperty.LabelContent, description: EKProperty.LabelContent, button: EKProperty.ButtonContent, action: @escaping EKPopUpMessageAction)
+```
+
 ## 0.2.4
 
 ### Features